add unit test for start_basic, move to own file
