+++ pkg/controller/elasticsearch/client/client_test.go
@@ -16,7 +16,6 @@ import (
 	"sort"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/elastic/cloud-on-k8s/pkg/controller/common/certificates"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/common/version"
@@ -438,91 +437,6 @@ func TestClient_Equal(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			require.True(t, tt.c1.Equal(tt.c2) == tt.want)
 		})
-	}
-}
-
-func TestClient_UpdateLicense(t *testing.T) {
-	tests := []struct {
-		expectedPath string
-		version      version.Version
-	}{
-		{
-			expectedPath: "/_xpack/license",
-			version:      version.MustParse("6.8.0"),
-		},
-		{
-			expectedPath: "/_license",
-			version:      version.MustParse("7.0.0"),
-		},
-	}
-	for _, tt := range tests {
-		testClient := NewMockClient(tt.version, func(req *http.Request) *http.Response {
-			require.Equal(t, tt.expectedPath, req.URL.Path)
-			return &http.Response{
-				StatusCode: 200,
-				Body:       ioutil.NopCloser(strings.NewReader(fixtures.LicenseUpdateResponseSample)),
-				Header:     make(http.Header),
-				Request:    req,
-			}
-		})
-		in := LicenseUpdateRequest{
-			Licenses: []License{
-				{
-					UID:                "893361dc-9749-4997-93cb-802e3d7fa4xx",
-					Type:               "basic",
-					IssueDateInMillis:  0,
-					ExpiryDateInMillis: 0,
-					MaxNodes:           1,
-					IssuedTo:           "unit-test",
-					Issuer:             "test-issuer",
-					Signature:          "xx",
-				},
-			},
-		}
-		got, err := testClient.UpdateLicense(context.Background(), in)
-		assert.NoError(t, err)
-		assert.Equal(t, true, got.Acknowledged)
-		assert.Equal(t, "valid", got.LicenseStatus)
-	}
-
-}
-
-func TestClient_GetLicense(t *testing.T) {
-	tests := []struct {
-		expectedPath string
-		version      version.Version
-	}{
-		{
-			expectedPath: "/_xpack/license",
-			version:      version.MustParse("6.8.0"),
-		},
-		{
-			expectedPath: "/_license",
-			version:      version.MustParse("7.0.0"),
-		},
-	}
-
-	for _, tt := range tests {
-		testClient := NewMockClient(tt.version, func(req *http.Request) *http.Response {
-			require.Equal(t, tt.expectedPath, req.URL.Path)
-			return &http.Response{
-				StatusCode: 200,
-				Body:       ioutil.NopCloser(strings.NewReader(fixtures.LicenseGetSample)),
-				Header:     make(http.Header),
-				Request:    req,
-			}
-		})
-		got, err := testClient.GetLicense(context.Background())
-		assert.NoError(t, err)
-		assert.Equal(t, "893361dc-9749-4997-93cb-802e3d7fa4xx", got.UID)
-		assert.Equal(t, "platinum", got.Type)
-		assert.EqualValues(t, time.Unix(0, 1548115200000*int64(time.Millisecond)).UTC(), *got.IssueDate)
-		assert.Equal(t, int64(1548115200000), got.IssueDateInMillis)
-		assert.EqualValues(t, time.Unix(0, 1561247999999*int64(time.Millisecond)).UTC(), *got.ExpiryDate)
-		assert.Equal(t, int64(1561247999999), got.ExpiryDateInMillis)
-		assert.Equal(t, 100, got.MaxNodes)
-		assert.Equal(t, "issuer", got.Issuer)
-		assert.Equal(t, int64(1548115200000), got.StartDateInMillis)
 	}
 }
